+++ src/array/every.pipe.ts
@@ -1,12 +1,12 @@
 import { Pipe, PipeTransform  } from '@angular/core';
-import { every } from '../utils/utils';
+import { every, CollectionPredicate } from '../utils/utils';
 
 @Pipe({
   name: 'every'
 })
 export class EveryPipe implements PipeTransform {
   
-  transform (input: any, predicate: Function): any {
+  transform (input: any, predicate: CollectionPredicate): any {
     
     return every(input, predicate);
   }